requirements.txt
pymongo==4.0.1
discord.py==1.7.3
discord-py-interactions==4.0.2